Седьмой спринт. Перед третьей отправкой на ревью
